Change types INTEGER -> LONG

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {100, 50, 70, 90};
+        long[] sales = {100, 50, 70, 90};
         SalesManager salesManager = new SalesManager(sales);
         System.out.println(salesManager.max());
     }
